+++ src/commands/lewd/e621.js
@@ -113,7 +113,7 @@ function tags(client, evt, suffix) {
               }
               file = null;
               let blacklistMatch = getOne(blacklist, tags);
-              imageDescription = `**BLACKLISTED RESULT** - Link Contains: ${blacklistMatch} | **Link:** [Click Here](https://e621.net/post/show/${id})`;
+              imageDescription = `**BLACKLISTED** - Matched: \`${blacklistMatch}\` | **Link:** [Click Here](https://e621.net/post/show/${id})`;
             }
           }