+++ src/commands/lewd/konachan.js
@@ -114,7 +114,7 @@ function tags(client, evt, suffix) {
               }
               fileurl = null;
               let blacklistMatch = getOne(blacklist, tags);
-              imageDescription = `**BLACKLISTED RESULT** - Link Contains: ${blacklistMatch} | **Link:** [Click Here](http://konachan.com/post/show/${id})`;
+              imageDescription = `**BLACKLISTED** - Matched: \`${blacklistMatch}\` | **Link:** [Click Here](http://konachan.com/post/show/${id})`;
             }
           }